Add auth redirect mechanism
When someone accepts a timesheet invite they might not have an account
yet, so we'll need to allow them to step through the auth flow and
remember where they came from.

Some OAuth providers don't allow for additional query params in the
redirect_uri or will error based on mismatches. This approach uses
short-lived cookies to work around this.