Allow unlinking of shared timesheets
Allows users, who have had a timesheet shared with them, to remove their
own access from it.

Abstracts the Danger Zone because depending on the share status,
different UI should be displayed (access control is managed in PB).

Adds a client-side auth helper just to help get the user id and email.
This will come from the layout to ensure that it's refreshed if the
cookie updates.