Make Hours/Target hours reusable
They previously had separate versions that had a lot of duplication
between them, when all they did differently was figure out when a week
started and when it ended.

This refactor means that TargetHours is always displayed if the timesheet
has a target, and that it's accessible via HoursWorked only

HoursWorked will now work without being placed in a given day's context,
and will accept its start + end dates as component params. This also
creates a more explicit declaration of which time window a certain
instance of this component covers